Update e.py
Changed int datatype to float datatype.

+++ e.py
@@ -6,8 +6,8 @@ def bill(a,b,c):
 	bill requries 3 parameters.
 	bill(<quantity of product>,<price of product>,<discount on product>)
 	"""
-	amount_with_quantity = int(b)*int(a)
-	calculating_discount = amount_with_quantity*int(c)/100
+	amount_with_quantity = float(b)*float(a)
+	calculating_discount = amount_with_quantity*float(c)/100
 	total_amount = amount_with_quantity - calculating_discount
 	t = total_amount
 	return t
